Major fixes

+++ src/app/admin/table/table.component.ts
@@ -3,7 +3,9 @@ import { DataService } from 'src/app/services/database';
 import { ConfirmationService, MessageService } from 'primeng/api';
 import { NgForm } from '@angular/forms';
 import { v4 as uuidv4 } from 'uuid';
-import { Subscription, forkJoin, of } from 'rxjs';
+import { Observable, Subscription, forkJoin, switchMap } from 'rxjs';
+import { NgxImageCompressService } from 'ngx-image-compress';
+import { NetworkStatusService } from 'src/app/services/networkstatus.service';
 @Component({
   selector: 'app-table',
   templateUrl: './table.component.html',
@@ -28,8 +30,6 @@ export class TableComponent implements OnInit {
     { name: '1 Year', code: 12 }
   ];
   removeImageHide: boolean = false;
-  pendingFeeUsers: any;
-
   feeDate: Date | undefined;
   dojDate: Date | undefined;
   uid: any;
@@ -48,7 +48,7 @@ export class TableComponent implements OnInit {
 
   dateTime = new Date();
 
-  constructor(private dataService: DataService, private confirmationService: ConfirmationService, private messageService: MessageService) {
+  constructor(private imageCompress: NgxImageCompressService, private dataService: DataService, private confirmationService: ConfirmationService, private messageService: MessageService, private networkStatusService: NetworkStatusService) {
     this.dateTime.setDate(this.dateTime.getDate());
   }
 
@@ -67,27 +67,60 @@ export class TableComponent implements OnInit {
     }
   }
 
-  ngOnInit() {
-    this.reset()
+  isOnline(): Observable<boolean> {
+    return this.networkStatusService.isOnline
   }
 
-  removeFile(uid) {
-    this.removeImage = this.dataService.removeImage(uid).subscribe({
-      next: (res) => {
-        this.removeImageHide = false
-        console.log('Removing Image Sucessfull');
-      },
-      error: (error) => {
-        console.log(error, "Remove Image Failed Error");
-        alert('Removing Image failed')
-      }
-    })
+  ngOnInit(): void {
+    this.reset();
   }
 
-  edit(product, pendingMode?) {
+  reset(): void {
+    this.showLoader = true
+    this.loading = true;
+    this.getData = this.isOnline().pipe(
+      switchMap(isOnline => {
+        if (!isOnline) {
+          this.messageService.add({ severity: 'error', summary: 'Network Error', detail: 'Internet connection lost' });
+          this.showLoader = false;
+          return [];
+        } else {
+          return this.dataService.getData();
+        }
+      })
+    ).
+      subscribe({
+        next: (data) => {
+          this.users = data;
+          this.users.forEach(ele => {
+            ele['package'] = this.packageId.find((id => id['code'] == ele['package']))['name'];
+          });
+          if (!this.userTableHdn) {
+            this.users = this.users.filter(object => object.culprit);
+          }
+          this.loading = false;
+          this.showLoader = false;
+        },
+        error: (error) => {
+          this.showLoader = false;
+          console.log(error, "Fetch Data failed");
+          alert('Getting issues while fetching data.')
+        }
+      });
+  }
+
+  removeFile() {
+    this.removeImageHide = false;
+    this.selectedFile = null;
+    if (this.fileUpload) {
+      this.fileUpload.clear();
+    }
+  }
+
+  edit(product) {
     this.feeDate = new Date(this.convertToDate(product['feedate']))
     this.dojDate = new Date(this.convertToDate(product['doj']))
-    this.editMode = pendingMode;
+    this.editMode = !this.userTableHdn;
 
     product = { ...product, package: this.packageId.find((id => id['name'] == product['package'])) };
     this.userdata = product
@@ -95,7 +128,9 @@ export class TableComponent implements OnInit {
     this.removeImageHide = true;
     this.visible = true;
     this.selectedFile = null;
-    this.fileUpload.clear();
+    if (this.fileUpload) {
+      this.fileUpload.clear();
+    }
   }
 
   delete(event: Event, uid) {
@@ -111,22 +146,34 @@ export class TableComponent implements OnInit {
 
       accept: () => {
         this.showLoader = true
-        this.forkJoin = forkJoin({
-          deleteData: this.dataService.deleteData(uid),
-          removeImage: this.dataService.removeImage(uid)
-        }).subscribe({
-          next: (results) => {
-            this.messageService.add({ severity: 'error', summary: 'Confirmed', detail: 'Record deleted' });
-          },
-          error: (error) => {
-            this.showLoader = false;
-            console.error('Operation failed', error);
-            alert('Operation failed');
-          },
-          complete: () => {
-            this.showLoader = false;
-          }
-        });
+
+        this.forkJoin = this.isOnline().pipe(
+          switchMap(isOnline => {
+            if (!isOnline) {
+              this.messageService.add({ severity: 'error', summary: 'Network Error', detail: 'Internet connection lost' });
+              this.showLoader = false;
+              return [];
+            } else {
+              return forkJoin({
+                deleteData: this.dataService.deleteData(uid),
+                removeImage: this.dataService.removeImage(uid)
+              })
+            }
+          })
+        )
+          .subscribe({
+            next: (results) => {
+              this.messageService.add({ severity: 'error', summary: 'Confirmed', detail: 'Record deleted' });
+            },
+            error: (error) => {
+              this.showLoader = false;
+              console.error('Operation failed', error);
+              alert('Operation failed');
+            },
+            complete: () => {
+              this.showLoader = false;
+            }
+          });
       },
       reject: () => {
         this.showLoader = false;
@@ -160,28 +207,6 @@ export class TableComponent implements OnInit {
     return this.users ? this.first === 0 : true;
   }
 
-  reset() {
-    this.showLoader = true
-    this.loading = true;
-    this.getData = this.dataService.getData()
-      .subscribe({
-        next: (data) => {
-          this.users = data;
-          this.users.forEach(ele => {
-            ele['package'] = this.packageId.find((id => id['code'] == ele['package']))['name'];
-          });
-          this.pendingFeeUsers = this.users.filter(object => object.culprit);
-          this.loading = false;
-          this.showLoader = false;
-        },
-        error: (error) => {
-          this.showLoader = false;
-          console.log(error, "Fetch Data failed");
-          alert('Getting issues while fetching data.')
-        }
-      });
-  }
-
   newUser() {
     this.userdata = {};
     this.visible = true;
@@ -192,14 +217,18 @@ export class TableComponent implements OnInit {
     this.selectedFile = null;
     this.uid = undefined;
     this.editMode = false;
-    this.fileUpload.clear();
+    if (this.fileUpload) {
+      this.fileUpload.clear();
+    }
   }
 
   cancel() {
     this.visible = false
     this.selectedFile = null;
     this.userForm.resetForm();
-    this.fileUpload.clear();
+    if (this.fileUpload) {
+      this.fileUpload.clear();
+    }
   }
 
   convertToDate(inputDateString) {
@@ -225,8 +254,46 @@ export class TableComponent implements OnInit {
     return `${day}/${month}/${year}`;
   }
 
-  getUniqueImageUrl(uid: string): string {
-    return `https://firebasestorage.googleapis.com/v0/b/srm-royal-gym.appspot.com/o/${uid}?alt=media&amp;token=70e5f4a4-0df3-4679-80af-1d981835a671&amp`;
+  compressFile(image: File, uid) {
+    const reader = new FileReader();
+    reader.readAsDataURL(image);
+    reader.onload = (event) => {
+      const imageDataUrl = event.target.result as string;
+
+      this.compressImage(imageDataUrl, image.name, uid);
+    };
+  }
+
+  compressImage(imageDataUrl: string, fileName: string, uid) {
+    const quality = 50;
+    const maxWidth = 800;
+    const maxHeight = 800;
+
+    this.imageCompress.compressFile(imageDataUrl, -1, quality, quality, maxWidth, maxHeight).then(
+      async (result: string) => {
+        let compressedFile = this.dataURLtoFile(result, fileName);
+        while (compressedFile.size > 100 * 1024) {
+          result = await this.imageCompress.compressFile(result, -1, quality / 2, quality / 2, maxWidth, maxHeight);
+          compressedFile = this.dataURLtoFile(result, fileName);
+        }
+        this.selectedFile = compressedFile
+        if (!this.imageUploaded && this.selectedFile != null) {
+          this.dataService.uploadImage(this.selectedFile, uid)
+        }
+      }
+    );
+  }
+
+  dataURLtoFile(dataurl: string, filename: string) {
+    const arr = dataurl.split(',');
+    const mime = arr[0].match(/:(.*?);/)[1];
+    const bstr = atob(arr[1]);
+    let n = bstr.length;
+    const u8arr = new Uint8Array(n);
+    while (n--) {
+      u8arr[n] = bstr.charCodeAt(n);
+    }
+    return new File([u8arr], filename, { type: mime });
   }
 
   addUser(product) {
@@ -235,38 +302,52 @@ export class TableComponent implements OnInit {
       alert('Please fill the form completly.')
       return;
     }
-    else {
-      if (this.userdata['uid']) {
-        this.uid = this.userdata['uid']
-      } else {
-        if (this.selectedFile == null) {
-          this.showLoader = false;
-          alert("Please select image.")
-          return;
-        } else {
-          this.uid = uuidv4()
-        }
+    else if (this.selectedFile == null && !this.removeImageHide) {
+      this.showLoader = false;
+      alert("Please select image.")
+      return;
+    }
+    if (this.userdata['uid']) {
+      this.uid = this.userdata['uid']
+    } else {
+      this.uid = uuidv4()
+    }
+
+    if (this.uid) {
+      this.showLoader = true;
+
+      product['package'] = product['package']['code'];
+
+      const feeDateObject = new Date(this.feeDate);
+      const dojDataObject = new Date(this.dojDate);
+      const nextDateObject = new Date(this.feeDate);
+      nextDateObject.setMonth(feeDateObject.getMonth() + product.package);
+
+      const timestamp = new Date().getTime();
+      let photoUrl = 'https://firebasestorage.googleapis.com/v0/b/srm-royal-gym.appspot.com/o/' + this.uid + '?alt=media&amp;token=70e5f4a4-0df3-4679-80af-1d981835a671&amp;t=' + timestamp
+
+      product = {
+        ...product, uid: this.uid, photo: photoUrl, nextFeeDate: this.convertToStrDate(nextDateObject),
+        doj: this.convertToStrDate(dojDataObject), feedate: this.convertToStrDate(feeDateObject),
+        culprit: this.isPastOrToday(this.convertToStrDate(nextDateObject))
       }
 
-      if (this.uid) {
-        this.showLoader = true;
+      if (this.selectedFile != null && !this.imageUploaded) {
+        this.compressFile(this.selectedFile, this.uid)
+      }
 
-        product['package'] = product['package']['code'];
-
-        const feeDateObject = new Date(this.feeDate);
-        const dojDataObject = new Date(this.dojDate);
-        const nextDateObject = new Date(this.feeDate);
-        nextDateObject.setMonth(feeDateObject.getMonth() + product.package);
-
-        product = {
-          ...product, uid: this.uid, nextFeeDate: this.convertToStrDate(nextDateObject),
-          doj: this.convertToStrDate(dojDataObject), feedate: this.convertToStrDate(feeDateObject),
-          culprit: this.isPastOrToday(this.convertToStrDate(nextDateObject))
-        }
-        this.forkJoinAddUser = forkJoin({
-          uploadImage: !this.imageUploaded ? (this.selectedFile != null ? this.dataService.uploadImage(this.selectedFile, this.uid) : of({})) : of({}),
-          setData: this.dataService.setData(this.uid, product)
-        })
+      setTimeout(() => {
+        this.forkJoinAddUser = this.isOnline().pipe(
+          switchMap(isOnline => {
+            if (!isOnline) {
+              this.messageService.add({ severity: 'error', summary: 'Network Error', detail: 'Internet connection lost' });
+              this.showLoader = false;
+              return [];
+            } else {
+              return this.dataService.setData(this.uid, product)
+            }
+          })
+        )
           .subscribe({
             next: (results) => {
               console.log(results, "Manish");
@@ -288,7 +369,9 @@ export class TableComponent implements OnInit {
               }
               this.selectedFile = null;
               this.visible = false;
-              this.fileUpload.clear();
+              if (this.fileUpload) {
+                this.fileUpload.clear();
+              }
               this.userForm.resetForm();
               this.showLoader = false;
             },
@@ -298,16 +381,19 @@ export class TableComponent implements OnInit {
               alert('Client add failed!');
             }
           });
-      }
+      }, 2000);
+
     }
   }
 
   allUser() {
     this.userTableHdn = true;
+    this.reset()
   }
 
   feePendindUser() {
     this.userTableHdn = false;
+    this.users = this.users.filter(object => object.culprit);
   }
 
   onFileSelect(event: any, fileUpload) {
